refactor(signer): moves signer to utils
